Adding logic for dst MAC address ExactMatch.

+++ conf/utils.py
@@ -67,7 +67,7 @@ def mac_by_interface(name):
 
 
 def mac2hex(mac):
-    return int(mac.replace(':', ''), 16)
+    return long(mac.replace(':', ''), 16)
 
 
 def peer_by_interface(name):
